Start cleanup by removing menu
There were just too many issues when using miru.ext.Menu.
This commit switch to using crescent directly to send
board initially, hikari to update it and only using a
miru view to display the buttons.

Moved the grid to a class to make referring to it and
its size easier as well as using grid setup functions