Add row and column labels to grid
Required lots of trial and error to find the correct mode(normal, code block),
chars and indentation but it does work for 1x1 through 99x99 but even discord
will not go that far.

Also switched back to the bomb char but now using a different version of it
which I think looks nicer.