Start adding dectalk vc plugin
Always uses test.wav but adding dectalk to generate new
audio files should not be too hard at this point.
There is also an issue with ongaku which means the first
call to player.connect to fail which caused a bit of a
mess in the code to reasonably report to the user.