Restart implementing minesweeper
New method reverts to using a menu but with a single
screen that just directly adds button to the menu.
This was done because clicking buttons in views is
much slower with there being a second waiting period
once the new buttons appear but before another input
is possible which does not happen with menus.
Due to the letter and number buttons being dynamic,
there is not much to be gained by having multiple
screens so just editing the menu is simpiler and
possibly faster.