Mostly fix issues with replies in checkers
I was initially using make_move to validate moves,
but it assumes that the token can move to the target
so I switched to checking that manually but kept
using the now removed return type of make_move.

Fix replies not allowing captures.
I am still not happy with needing to check both
cases but it at least works now.